Bonus 2 - DictComprehension
If headers are presented, then we need to initialize the dictionary by
Creating a list with a single entry consisting of the header.
If headers are not presented, then we don't initialize anything
And let defaultdict handle the missing key issues.
We set headers to reader.fieldnames (note - just a property)
And then we are able to assign the value by using the zip
of  the headers and the values for each row.