Second Version -  using zip, function list.
I always feel a bit dirty whenever I use an index - almost all of the
time I probably can get by with a zip function.  Trey pointed out
the right path.  Also - learned a clever new trick  - you can do
method(var, params) instead of var.method(params) for string
functions.  Which lets you build up a list of functions.

Also - learned (or rediscovered) that python has a built
in ljust, rjust to go with it's ltrim, rtrim functions.

So - Double use of zip in this exercise - first to break out the rows
so I can see them in colums, and iterate the max lengths,
the second was to zip in the widths and alignments into
the parameter list without having to use indexes.