deep_flatten.py Pretty straightforward - had to look up where Iterable was defined - it's not a builtin - so, in the regular place - collections.abc.Iterable.
Thrrew in a check to see if something  type(seq) == str,
I'll be interested to see how far off from Trey's solution I am.

It's recursive - what's interesting is that  the loop:
            if isinstance(item,Iterable):
                for element in deep_flatten(item):
                    yield element

does what I would hope it would do - the yields pass all
the way back to the top of the call chain.