all_same I actually ended up being able to solve this with sets by using "repr" to force unhashable items into a hasheable format. What Trey was really interested in for this problem though was for people to struggle with doing a large number of logical tests, and then discover two features: 1) pulling the first element of an iterable off with:    next(iter(iterable), None) (The "None" prevents a stop iteration error)
2) The use of "all" - I think that's what the real focus was here.
all provides a very fast, and pythonic way to test a large number
of logical tests, and exits as soon as one of the tests is false.

A nice test of this is:
def is_even(n):
    print (f'{n} even: {not n % 2}')
    return not n % 2
lst=[2,4,6,200,5,50]
print( all(is_even(x) for x in lst))
....
2 even: True
4 even: True
6 even: True
200 even: True
5 even: False
False

So you can see that after 5 is tested, you get an immediate exit.

Two things that I took away from this -
1) always feel comfortable using next(iter(iterable),None)  - that
  is a pythonic and fine way to pull the first element of an iterable.

2) wherever possible, see if you can use any "all" or "any" to
    test a large number of logical tests - with the bonus
    of immediate exit if anything goes false.