Expiring Generators + namedTuples. Took me 5 or so minutes to realize you can't reference a generator twice on the same line.  That's a problem, this should be laser clear to me by now.  I really need to start rewiring my brain whenever I see an Iterator, or generator, or the possibility that the iterable may be one of those - and instantly create a ticking time bomb, or water dripping out of a container, or *something* that makes me realize we may face a StopIteration error, or, may silently end up with the empty iterator on the second run.
But - empty iterator *IS STILL TRUTHY* - its different from None.
You can't do:
   while i:
     print(next(i))
Because you will exit with a StopIteration error.

Anyways - that was the bonus I stumbled over.  The NamedTuple I
nailed.  Well, relied on intellisense for the case of namedtuple.
But I tink I've finally got my head wrapped around the concept
of Factory versus instance.  Still not sure what the purpose
of the "name" in the first part is.  Maybe to make it look
good on REPRs I guess.