Multiple Keys - keyword only parser
Python has this weird *search keyword only parser feature that landed
back in the early days of python 3.0.   It's a way to have a single
parameter that can have mutiple arguments - landing in *search
in this case.   I kludge a solution here by checking to see if
it has just one key - in which case I convert from a tuple ->
string, and then continue on the normal path *else*
I return a TUPLE of a GENERATIOR function which
combines all of the elements of the tuple multi
argument parameter.  Horrible.